# sourcode/golang/server-microservices/modules/cart/endpoints/api_shopping_cart.go
/*
 * Swagger Petstore - OpenAPI 3.0
 *
 * This is a example for performance study based on the OpenAPI 3.0 specification.
 *
 * API version: 1.0.11
 * Contact: skalicky.martin@iotdomu.cz
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package endpoints

import (
	"microservices/modules/cart/middleware"
	"microservices/shared/getters"
	"net/http"

	"github.com/gorilla/mux"
)

func AddItemToCart(w http.ResponseWriter, r *http.Request) {
	var ctx = r.Context()
	var db = middleware.GetDb(ctx)
	var itemApi = middleware.GetItemApi(ctx)
	var userId = getters.GetUserId(r)
	var params = mux.Vars(r)

	var item, err = itemApi.GetItemByItemId(ctx, params["itemId"])
	if err != nil {
		failUnexpectedError(err, w, r)
		return
	}

	cart, err := db.GetCartByUserId(ctx, userId)
	if err != nil {
		failUnexpectedError(err, w, r)
		return
	}

	_, err = db.AddItemToCart(ctx, item.Id, cart)
	if err != nil {
		failUnexpectedError(err, w, r)
		return
	}

	itemsIds, err := db.GetItemsIdsInCart(ctx, cart)
	if err != nil {
		failUnexpectedError(err, w, r)
		return
	}

	items, err := itemApi.GetItemsByIds(ctx, itemsIds)
	if err != nil {
		failUnexpectedError(err, w, r)
		return
	}

	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	jsonResponse(items, w)
}

func RemoveItemFromCart(w http.ResponseWriter, r *http.Request) {
	var ctx = r.Context()
	var db = middleware.GetDb(ctx)
	var itemApi = middleware.GetItemApi(ctx)
	var userId = getters.GetUserId(r)
	var params = mux.Vars(r)

	var item, err = itemApi.GetItemByItemId(ctx, params["itemId"])
	if err != nil {
		failUnexpectedError(err, w, r)
		return
	}

	cart, err := db.GetCartByUserId(ctx, userId)
	if err != nil {
		failUnexpectedError(err, w, r)
		return
	}

	_, err = db.RemoveItemFromCart(ctx, item.Id, cart)
	if err != nil {
		failUnexpectedError(err, w, r)
		return
	}

	itemsIds, err := db.GetItemsIdsInCart(ctx, cart)
	if err != nil {
		failUnexpectedError(err, w, r)
		return
	}

	items, err := itemApi.GetItemsByIds(ctx, itemsIds)
	if err != nil {
		failUnexpectedError(err, w, r)
		return
	}

	w.Header().Set("Content-Type", "application/json; charset=UTF-8")
	jsonResponse(items, w)
}